Consolidate all qunit testing into test_qunit.py.
